refactor: internalize selected map loader's planning helper
  * Also merge selected tests, and move selected node test to pointcloud node test
