+++ map/autoware_map_loader/CMakeLists.txt
@@ -71,8 +71,7 @@ if(BUILD_TESTING)
   add_testcase(test/pointcloud_map_loader/test_partial_map_loader_core.cpp)
   add_testcase(test/pointcloud_map_loader/test_differential_map_loader_module.cpp)
   add_testcase(test/pointcloud_map_loader/test_differential_map_loader_core.cpp)
-  add_testcase(test/pointcloud_map_loader/test_selected_map_loader_module.cpp)
-  add_testcase(test/pointcloud_map_loader/test_selected_map_loader_core.cpp)
+  add_testcase(test/pointcloud_map_loader/test_selected_map_loader.cpp)
   add_testcase(test/lanelet2_map_loader/test_lanelet2_map_loader_utils.cpp)
   add_testcase(test/lanelet2_map_loader/test_lanelet2_map_cell_metadata.cpp)
   add_testcase(test/lanelet2_map_loader/test_lanelet2_selected_map_loader_module.cpp)
