+++ map/autoware_map_loader/test/pointcloud_map_loader/test_selected_map_loader.cpp
@@ -1,0 +1,169 @@
+// Copyright 2026 The Autoware Contributors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "../src/pointcloud_map_loader/selected_map_loader.hpp"
+
+#include <gtest/gtest.h>
+#include <pcl/io/pcd_io.h>
+#include <pcl/point_types.h>
+
+#include <map>
+#include <memory>
+#include <string>
+#include <vector>
+
+namespace autoware::map_loader
+{
+namespace
+{
+using GetSelectedPointCloudMap = autoware_map_msgs::srv::GetSelectedPointCloudMap;
+
+PCDFileMetadata make_metadata(
+  const float min_x, const float min_y, const float min_z, const float max_x, const float max_y,
+  const float max_z)
+{
+  PCDFileMetadata metadata;
+  metadata.min.x = min_x;
+  metadata.min.y = min_y;
+  metadata.min.z = min_z;
+  metadata.max.x = max_x;
+  metadata.max.y = max_y;
+  metadata.max.z = max_z;
+  return metadata;
+}
+
+void write_dummy_pcd(const std::string & path)
+{
+  pcl::PointCloud<pcl::PointXYZ> cloud;
+  cloud.width = 1;
+  cloud.height = 1;
+  cloud.is_dense = true;
+  cloud.points.resize(1);
+  cloud.points[0] = pcl::PointXYZ(0.0F, 0.0F, 0.0F);
+  pcl::io::savePCDFileASCII(path, cloud);
+}
+
+TEST(SelectedMapLoader, ReturnsLoadAndMissingIdsInRequestOrder)
+{
+  const std::string a_path = "/tmp/test_selected_map_loader_a.pcd";
+  const std::string b_path = "/tmp/test_selected_map_loader_b.pcd";
+  write_dummy_pcd(a_path);
+  write_dummy_pcd(b_path);
+
+  const std::map<std::string, PCDFileMetadata> metadata_dict = {
+    {a_path, make_metadata(-1.0F, -1.0F, -1.0F, 1.0F, 1.0F, 1.0F)},
+    {b_path, make_metadata(-2.0F, -2.0F, -1.0F, 2.0F, 2.0F, 1.0F)},
+  };
+
+  std::vector<std::string> errors;
+  SelectedMapLoaderModule module(metadata_dict, [&errors](const std::string & msg) {
+    errors.push_back(msg);
+  });
+
+  auto request = std::make_shared<GetSelectedPointCloudMap::Request>();
+  request->cell_ids = {"missing_0.pcd", b_path, "missing_1.pcd", a_path};
+  auto response = std::make_shared<GetSelectedPointCloudMap::Response>();
+
+  EXPECT_TRUE(module.create_response(request, response));
+
+  ASSERT_EQ(response->new_pointcloud_with_ids.size(), 2U);
+  EXPECT_EQ(response->new_pointcloud_with_ids[0].cell_id, b_path);
+  EXPECT_EQ(response->new_pointcloud_with_ids[1].cell_id, a_path);
+  ASSERT_EQ(errors.size(), 2U);
+  EXPECT_FALSE(errors[0].empty());
+  EXPECT_FALSE(errors[1].empty());
+  EXPECT_EQ(response->header.frame_id, "map");
+}
+
+TEST(SelectedMapLoader, EmptyRequestProducesEmptyResponse)
+{
+  const std::string a_path = "/tmp/test_selected_map_loader_empty_a.pcd";
+  write_dummy_pcd(a_path);
+
+  const std::map<std::string, PCDFileMetadata> metadata_dict = {
+    {a_path, make_metadata(-1.0F, -1.0F, -1.0F, 1.0F, 1.0F, 1.0F)},
+  };
+
+  SelectedMapLoaderModule module(metadata_dict);
+
+  auto request = std::make_shared<GetSelectedPointCloudMap::Request>();
+  auto response = std::make_shared<GetSelectedPointCloudMap::Response>();
+
+  EXPECT_TRUE(module.create_response(request, response));
+
+  EXPECT_TRUE(response->new_pointcloud_with_ids.empty());
+  EXPECT_EQ(response->header.frame_id, "map");
+}
+
+TEST(SelectedMapLoader, DuplicateRequestsArePreserved)
+{
+  const std::string a_path = "/tmp/test_selected_map_loader_duplicate_a.pcd";
+  write_dummy_pcd(a_path);
+
+  const std::map<std::string, PCDFileMetadata> metadata_dict = {
+    {a_path, make_metadata(-1.0F, -1.0F, -1.0F, 1.0F, 1.0F, 1.0F)},
+  };
+
+  std::vector<std::string> errors;
+  SelectedMapLoaderModule module(metadata_dict, [&errors](const std::string & msg) {
+    errors.push_back(msg);
+  });
+
+  auto request = std::make_shared<GetSelectedPointCloudMap::Request>();
+  request->cell_ids = {a_path, a_path, "missing.pcd", "missing.pcd"};
+  auto response = std::make_shared<GetSelectedPointCloudMap::Response>();
+
+  EXPECT_TRUE(module.create_response(request, response));
+
+  ASSERT_EQ(response->new_pointcloud_with_ids.size(), 2U);
+  EXPECT_EQ(response->new_pointcloud_with_ids[0].cell_id, a_path);
+  EXPECT_EQ(response->new_pointcloud_with_ids[1].cell_id, a_path);
+  ASSERT_EQ(errors.size(), 2U);
+  EXPECT_FALSE(errors[0].empty());
+  EXPECT_FALSE(errors[1].empty());
+  EXPECT_EQ(response->header.frame_id, "map");
+}
+
+TEST(SelectedMapLoader, CreateMetadataSetsMapFrameAndCellBounds)
+{
+  const std::map<std::string, PCDFileMetadata> metadata_dict = {
+    {"a.pcd", make_metadata(0.0F, 1.0F, 2.0F, 10.0F, 11.0F, 12.0F)},
+    {"b.pcd", make_metadata(-5.0F, -6.0F, -7.0F, 5.0F, 6.0F, 7.0F)},
+  };
+
+  const auto msg = create_metadata(metadata_dict);
+
+  EXPECT_EQ(msg.header.frame_id, "map");
+  ASSERT_EQ(msg.metadata_list.size(), 2U);
+
+  EXPECT_EQ(msg.metadata_list[0].cell_id, "a.pcd");
+  EXPECT_FLOAT_EQ(msg.metadata_list[0].metadata.min_x, 0.0F);
+  EXPECT_FLOAT_EQ(msg.metadata_list[0].metadata.min_y, 1.0F);
+  EXPECT_FLOAT_EQ(msg.metadata_list[0].metadata.max_x, 10.0F);
+  EXPECT_FLOAT_EQ(msg.metadata_list[0].metadata.max_y, 11.0F);
+
+  EXPECT_EQ(msg.metadata_list[1].cell_id, "b.pcd");
+  EXPECT_FLOAT_EQ(msg.metadata_list[1].metadata.min_x, -5.0F);
+  EXPECT_FLOAT_EQ(msg.metadata_list[1].metadata.min_y, -6.0F);
+  EXPECT_FLOAT_EQ(msg.metadata_list[1].metadata.max_x, 5.0F);
+  EXPECT_FLOAT_EQ(msg.metadata_list[1].metadata.max_y, 6.0F);
+}
+}  // namespace
+}  // namespace autoware::map_loader
+
+int main(int argc, char ** argv)
+{
+  testing::InitGoogleTest(&argc, argv);
+  return RUN_ALL_TESTS();
+}
